requirements.txt
# conda env create --file <this_file.txt>
python=3.10
numpy=1.23
pandas=1.5
matplotlib=3.6
pvlib-python=0.9